+++ src/components/Dragonfly/Dragonfly.js
@@ -5,7 +5,7 @@ import './dragonfly.css';
 /**
  * Import actions here
  */
-import {getSelectedNode, getSinks} from '../../reducers/dragonflyReducer';
+import {getSelectedNode, getSinks, getSources} from '../../reducers/dragonflyReducer';
 
 const dragonfly = React.createClass({
     getInitialState: function(){
@@ -14,7 +14,7 @@ const dragonfly = React.createClass({
     },
     render: function(){
         return <div id="dragonfly">
-        <div id="left"></div>
+        <div id="left"><ul>{ this.props.sources.map(v => <li key={v.hash}>{v.hash}</li>)}</ul></div>
         <div id="center" key="2">{ this.props.selectedNode.hash }</div>
         <div id="right"><ul>{ this.props.sinks.map(v => <li key={v.hash}>{v.hash}</li>)}</ul></div>
         </div>
@@ -24,6 +24,7 @@ const dragonfly = React.createClass({
 const mapStateToProps = state => ({
     selectedNode: getSelectedNode(state),
     sinks: getSinks(state),
+    sources: getSources(state)
 });
 
 const mapDispatchToProps = dispatch => ({